+++ HISTORY.md
@@ -15,6 +15,7 @@ and this project adheres to `Semantic Versioning: https://semver.org/spec/v2.0.0
 * Support for selective access via range descriptor
 * Support for HLS authentication using HLS-GMAC.
 * Support for GlobalCiphering 
+* Parsing of ProfileGeneric buffer
 
 ### Changed